Interface

Transducer interface

Transducers.R_ — Type.

Transducers.R_{X}

When defining a transducer type X, it is often required to dispatch on type rf::R_{X} (Reducing Function) which bundles the current transducer rf.xform::X and the inner reducing function rf.inner::R_.

const R_{X} = Reduction{<:X}

source

Transducers.start — Function.

Transducers.start(rf::R_{X}, state)

This is an optional interface for a transducer. Default implementation just calls start of the inner reducing function; i.e.,

start(rf::Reduction, result) = start(rf.inner, result)

If the transducer X is stateful, it can "bundle" its private state with state (so that next function can be "pure").

start(rf::R_{X}, result) = wrap(rf, PRIVATE_STATE, start(rf.inner, result))

See Take, PartitionBy, etc. for real-world examples.

Side notes: There is no related API in Clojure's Transducers. Transducers.jl uses it to implement stateful transducers using "pure" functions. The idea is based on a slightly different approach taken in C++ Transducer library atria.

source

Transducers.next — Function.

Transducers.next(rf::R_{X}, state, input)

This is the only required interface. It takes the following form (if start is not defined):

next(rf::R_{X}, result, input) =
    # code calling next(rf.inner, result, possibly_modified_input)

See Map, Filter, Cat, etc. for real-world examples.

source

Transducers.complete — Function.

Transducers.complete(rf::R_{X}, state)

This is an optional interface for a transducer. If transducer X has some internal state, this is the last chance to "flush" the result.

See PartitionBy, etc. for real-world examples.

If both complete(rf::R_{X}, state) and start(rf::R_{X}, state) are defined, complete(rf::R_{X}, state) must unwarp state before returning state to the outer reducing function. If complete is not defined for R_{X}, this happens automatically.

source