Manual

Transducers and Transducer contexts

Base.append!

Base.collect

Base.copy!

Base.map!

Base.mapfoldl

Transducers.eduction

Transducers.loop

Transducers.transduce

Transducers.Cat

Transducers.Completing

Transducers.Dedupe

Transducers.Distinct

Transducers.Drop

Transducers.DropLast

Transducers.DropWhile

Transducers.Filter

Transducers.GetIndex

Transducers.Interpose

Transducers.Keep

Transducers.Map

Transducers.MapCat

Transducers.PartitionBy

Transducers.Replace

Transducers.Scan

Transducers.SetIndex

Transducers.Take

Transducers.TakeNth

Transducers.TakeWhile

Transducers.TeeZip

Transducers.Window

Transducer contexts

Base.mapfoldl — Function.

mapfoldl(xf, step, init, itr) :: T
transduce(xf, step, init, itr) :: Union{T, Reduced{T}}

Compose transducer xf with reducing step function step and reduce itr using it.

Warning

transduce differs from mapfoldl as Reduced{T} is returned if the transducer xf or step aborts the reduction.

This API is modeled after transduce in Clojure.

Arguments

xf::Transducer: A transducer.

step: A callable which accepts 1 and 2 arguments. If it only accepts 2 arguments, wrap it by Completing to add complete protocol.

init: An initial value fed to the first argument to reducing step function step.

itr: An iterable.

Examples

julia> using Transducers

julia> function step_demo(state, input)
           @show state, input
           state + input
       end;

julia> function step_demo(state)
           println("Finishing with state = ", state)
           state
       end;

julia> mapfoldl(Filter(isodd), step_demo, 0.0, 1:4)
(state, input) = (0.0, 1)
(state, input) = (1.0, 3)
Finishing with state = 4.0
4.0

source

Transducers.transduce — Function.

mapfoldl(xf, step, init, itr) :: T
transduce(xf, step, init, itr) :: Union{T, Reduced{T}}

See mapfoldl.

source

Transducers.loop — Function.

loop(step, xf, init, coll)

This is a shorthand for transduce(Completing(step), xf, init, coll). It is intended to be used with do block.

Examples

julia> using Transducers

julia> loop(Filter(isodd), 0.0, 1:4) do state, input
           @show state, input
           state + input
       end
(state, input) = (0.0, 1)
(state, input) = (1.0, 3)
4.0

source

Transducers.eduction — Function.

eduction(xf::Transducer, coll)

Create a iterable and reducible object.

Iterable.

Reducible (TODO); i.e., it can be handled by transduce efficiently.

This API is modeled after eduction in Clojure.

Examples

julia> using Transducers

julia> for x in eduction(Filter(isodd) |> Take(3), 1:1000)
           @show x
       end
x = 1
x = 3
x = 5

source

Base.map! — Function.

map!(xf::Transducer, dest, src)

Feed src to transducer xf, storing the result in dest. Collections dest and src must have the same shape. Transducer xf may contain filtering transducers. If some entries src are skipped, the corresponding entries in dest will be unchanged. Transducer xf must not contain any expansive transducers such as MapCat.

See also copy!.

Examples

julia> using Transducers

julia> xs = collect(1:5)
       ys = zero(xs)
       map!(Filter(isodd), ys, xs)
5-element Array{Int64,1}:
 1
 0
 3
 0
 5

julia> ans === ys
true

source

Base.copy! — Function.

copy!(xf::Transducer, dest, src)

Feed src to transducer xf, storing the result in dest. Collections dest and src may have the same shape. Source src must be iterable. Destination dest must implement empty! and push!.

See also map!.

Examples

julia> using Transducers

julia> copy!(PartitionBy(x -> x ÷ 3) |> Map(sum), Int[], 1:10)
4-element Array{Int64,1}:
  3
 12
 21
 19

source

Base.append! — Function.

append!(xf::Transducer, dest, src)

This API is modeled after into in Clojure.

Examples

julia> using Transducers

julia> append!(Drop(2), [-1, -2], 1:5)
5-element Array{Int64,1}:
 -1
 -2
  3
  4
  5

source

Base.collect — Function.

collect(xf::Transducer, itr)

Process an iterable itr using a transducer xf and collect the result into a Vector.

Examples

julia> using Transducers

julia> collect(Interpose(missing), 1:3)
5-element Array{Union{Missing, Int64},1}:
 1
  missing
 2
  missing
 3

source

Transducers

Transducers.Cat — Type.

Cat()

Examples

julia> using Transducers

julia> collect(Cat(), [[1, 2], [3], [4, 5]]) == 1:5
true

source

Transducers.Dedupe — Type.

Dedupe()

Examples

julia> using Transducers

julia> collect(Dedupe(), [1, 1, 2, 1, 3, 3, 2])
5-element Array{Int64,1}:
 1
 2
 1
 3
 2

source

Transducers.Distinct — Type.

Distinct()

Examples

julia> using Transducers

julia> collect(Distinct(), [1, 1, 2, 1, 3, 3, 2])
3-element Array{Int64,1}:
 1
 2
 3

source

Transducers.Drop — Type.

Drop(n)

Examples

julia> using Transducers

julia> collect(Drop(3), 1:5)
2-element Array{Int64,1}:
 4
 5

source

Transducers.DropLast — Type.

DropLast(n)

Examples

julia> using Transducers

julia> collect(DropLast(2), 1:5)
3-element Array{Int64,1}:
 1
 2
 3

source

Transducers.DropWhile — Type.

DropWhile(pred)

Examples

julia> using Transducers

julia> collect(DropWhile(x -> x < 3), 1:5)
3-element Array{Int64,1}:
 3
 4
 5

source

Transducers.Filter — Type.

Filter(pred)

Examples

julia> using Transducers

julia> collect(Filter(iseven), 1:3)
1-element Array{Int64,1}:
 2

source

Transducers.GetIndex — Type.

GetIndex(array)
GetIndex{inbounds}(array)

Examples

julia> using Transducers

julia> collect(GetIndex(1:10), [2, 3, 4])
3-element Array{Int64,1}:
 2
 3
 4

julia> collect(GetIndex{true}(1:10), [2, 3, 4])
3-element Array{Int64,1}:
 2
 3
 4

source

Transducers.Interpose — Type.

Interpose(sep)

Examples

julia> using Transducers

julia> collect(Interpose(missing), 1:3)
5-element Array{Union{Missing, Int64},1}:
 1
  missing
 2
  missing
 3

source

Transducers.Keep — Type.

Keep(f)

Examples

julia> using Transducers

julia> xf = Keep() do x
           if x < 3
               x + 1
           end
       end;

julia> collect(xf, 1:5)
2-element Array{Int64,1}:
 2
 3

source

Transducers.Map — Type.

Map(f)

Examples

julia> using Transducers

julia> collect(Map(x -> 2x), 1:3)
3-element Array{Int64,1}:
 2
 4
 6

source

Transducers.MapCat — Type.

MapCat(f)

Examples

julia> using Transducers

julia> collect(MapCat(x -> 1:x), 1:3)
6-element Array{Int64,1}:
 1
 1
 2
 1
 2
 3

source

Transducers.PartitionBy — Type.

PartitionBy(f)

Examples

julia> using Transducers

julia> collect(PartitionBy(x -> (x + 1) ÷ 3) |> Map(copy), 1:9)
4-element Array{Array{Int64,1},1}:
 [1]
 [2, 3, 4]
 [5, 6, 7]
 [8, 9]

source

Transducers.Replace — Type.

Replace(dict)

Examples

julia> using Transducers

julia> collect(Replace(Dict('a' => 'A')), "abc")
3-element Array{Char,1}:
 'A'
 'b'
 'c'

source

Transducers.Scan — Type.

Scan(f, [init])

Examples

julia> using Transducers

julia> collect(Scan(*), 1:3)
3-element Array{Int64,1}:
 1
 2
 6

julia> collect(Map(x -> x + im) |> Scan(*), 1:3)
3-element Array{Complex{Int64},1}:
 1 + 1im
 1 + 3im
 0 + 10im

julia> collect(Scan(*, 10), 1:3)
3-element Array{Int64,1}:
 10
 20
 60

source

Transducers.SetIndex — Type.

SetIndex(array)
SetIndex{inbounds}(array)

Examples

julia> using Transducers

julia> ys = zeros(3);

julia> mapfoldl(SetIndex(ys), first ∘ tuple, nothing, [(1, 11.1), (3, 33.3)])

julia> ys
3-element Array{Float64,1}:
 11.1
  0.0
 33.3

source

Transducers.Take — Type.

Take(n)

Examples

julia> using Transducers

julia> collect(Take(2), 1:10)
2-element Array{Int64,1}:
 1
 2

source

Transducers.TakeNth — Type.

TakeNth(n)

Examples

julia> using Transducers

julia> collect(TakeNth(3), 1:9)
3-element Array{Int64,1}:
 1
 4
 7

source

Transducers.TakeWhile — Type.

TakeWhile(pred)

Examples

julia> using Transducers

julia> collect(TakeWhile(x -> x < 3), 1:10)
2-element Array{Int64,1}:
 1
 2

source

Transducers.TeeZip — Type.

TeeZip(xform::Transducer)

Branch input into two "flows", inject one into xform and then merge the output of xform with the original input.

To illustrate how it works, consider the following usage

xf0 |> TeeZip(xf1) |> xf2

where xf0, xf1, and xf2 are some transducers. Schematically, the output yn from xfn flows as follows:

xf0      xf1                       xf2
---- y0 ------ y1 ---.-- (y0, y1) ----->
      |              |
       `-------------'
    "Tee"          "Zip"

Examples

julia> using Transducers

julia> collect(TeeZip(Filter(isodd) |> Map(x -> x + 1)), 1:5)
3-element Array{Tuple{Int64,Int64},1}:
 (1, 2)
 (3, 4)
 (5, 6)

source

Transducers.Window — Type.

Window(size, step = size, flush = false)
Window(size; step = size, flush = false)

Examples

julia> using Transducers

julia> collect(Window(3) |> Map(copy), 1:8)
2-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [4, 5, 6]

julia> collect(Window(3; flush=true) |> Map(copy), 1:8)
3-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [4, 5, 6]
 [7, 8]

julia> collect(Window(3; step=1) |> Map(copy), 1:8)
6-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [2, 3, 4]
 [3, 4, 5]
 [4, 5, 6]
 [5, 6, 7]
 [6, 7, 8]

source

Miscellaneous

Transducers.Completing — Type.

Completing(function)

Wrap a function to add a no-op complete protocol. Use it when passing a function without 1-argument arity to transduce etc.

source