Internals

Internals

Transducers.AdHocXF — Type.

AdHocXF(f, init, [onlast])

Examples

julia> using Transducers
       using Transducers: AdHocXF, @next
       using Setfield: @set!

julia> flushlast(rf, result) = rf(@next(rf, result, result.state));

julia> xf = AdHocXF(nothing, flushlast) do rf, result, input
           m = match(r"^name:(.*)", input)
           if m === nothing
               push!(result.state.lines, input)
               return result
           else
               chunk = result.state
               @set! result.state = (name=strip(m.captures[1]), lines=String[])
               push!(result.state.lines, input)
               if chunk === nothing
                   return result
               else
                   return rf(result, chunk)
               end
           end
       end;

julia> collect(xf, split("""
       name: Map
       type: onetoone
       name: Cat
       type: expansive
       name: Filter
       type: contractive
       name: Cat |> Filter
       type: chaotic
       """, "\n"; keepempty=false))
4-element Array{NamedTuple{(:name, :lines),Tuple{SubString{String},Array{String,1}}},1}:
 (name = "Map", lines = ["name: Map", "type: onetoone"])
 (name = "Cat", lines = ["name: Cat", "type: expansive"])
 (name = "Filter", lines = ["name: Filter", "type: contractive"])
 (name = "Cat |> Filter", lines = ["name: Cat |> Filter", "type: chaotic"])

source

Transducers.DefaultInit — Type.

DefaultInit(op)

DefaultInit is like InitialValues.Init but strictly internal to Transducers.jl. It is used for checking if the bottom reducing function is never called.

source

Transducers.UseSIMD — Type.

UseSIMD{ivdep}()

Tell the reducible to run the inner reducing function using @simd. The reducible can support it using @simd_if.

source

Transducers._foldl_blockarray — Method.

_foldl_blockarray(rf, acc, coll::BlockArrays.BlockArray)

source

Transducers._foldl_lazy_hcat — Method.

_foldl_lazy_hcat(rf, acc, coll::LazyArrays.Hcat)

source

Transducers._foldl_lazy_vcat — Method.

_foldl_lazy_vcat(rf, acc, coll::LazyArrays.Vcat)

source

Transducers._set_joiner_value — Method.

_set_joiner_value(ps::PrivateState, x) :: PrivateState

Set .state field of the PrivateState of the first "unbalanced" Joiner. A Joiner matched with preceding Splitter would be treated as a regular reducing function node. Thus, private state ps must have one more Joiner than Splitter.

source

Transducers.air — Method.

air.(broadcasting_expression) :: Broadcasted

Broadcast without materialization.

The idea is taken from @dawbarton's _lazy function: https://discourse.julialang.org/t/19641/20.

source

Transducers.check_no_ivdep — Method.

check_no_ivdep(::SIMDFlag)
check_no_ivdep(; simd=Val(false), kwargs...)

source

Transducers.foldl_nocomplete — Method.

foldl_nocomplete(rf, init, coll)

Call __foldl__ without calling complete.

source

Transducers.maybe_usesimd — Method.

maybe_usesimd(xform, simd)

Insert UseSIMD to xform if appropriate.

Arguments

xform::Transducer

simd: false, true, or :ivdep.

Examples

julia> using Transducers
       using Transducers: maybe_usesimd

julia> maybe_usesimd(reducingfunction(Map(identity), right), false)
Reduction{▶ NOTYPE}(
    Map(identity),
    BottomRF{▶ NOTYPE}(
        Transducers.right))

julia> maybe_usesimd(reducingfunction(Map(identity), right), true)
Reduction{▶ NOTYPE}(
    Transducers.UseSIMD{false}(),
    Reduction{▶ NOTYPE}(
        Map(identity),
        BottomRF{▶ NOTYPE}(
            Transducers.right)))

julia> maybe_usesimd(reducingfunction(Cat(), right), true)
Reduction{▶ NOTYPE}(
    Cat(),
    Reduction{▶ NOTYPE}(
        Transducers.UseSIMD{false}(),
        BottomRF{▶ NOTYPE}(
            Transducers.right)))

julia> maybe_usesimd(reducingfunction(Map(sin) |> Cat() |> Map(cos), right), :ivdep)
Reduction{▶ NOTYPE}(
    Map(sin),
    Reduction{▶ NOTYPE}(
        Cat(),
        Reduction{▶ NOTYPE}(
            Transducers.UseSIMD{true}(),
            Reduction{▶ NOTYPE}(
                Map(cos),
                BottomRF{▶ NOTYPE}(
                    Transducers.right)))))

julia> maybe_usesimd(
           reducingfunction(
               Map(sin) |> Cat() |> Map(cos) |> Cat() |> Map(tan),
               right,
           ),
           true,
       )
Reduction{▶ NOTYPE}(
    Map(sin),
    Reduction{▶ NOTYPE}(
        Cat(),
        Reduction{▶ NOTYPE}(
            Map(cos),
            Reduction{▶ NOTYPE}(
                Cat(),
                Reduction{▶ NOTYPE}(
                    Transducers.UseSIMD{false}(),
                    Reduction{▶ NOTYPE}(
                        Map(tan),
                        BottomRF{▶ NOTYPE}(
                            Transducers.right)))))))

source

Transducers.needintype — Method.

needintype(xf::Transducer) :: Bool
needintype(T::Type{<:Transducer}) :: Bool

Return false if start(xf::T, _) does not need intype. Abstract Transducer defaults to return true. This is because it is impossible to know if a user-defined transducer needs intype (typically via initvalue).

source

Transducers.reform — Method.

reform(rf, f)

Reset "bottom" reducing function of rf to f.

source

Transducers.simple_transduce — Method.

simple_transduce(xform, step, init, coll)

Simplified version of transduce. For simple transducers Julia may be able to emit a good code. This function exists only for performance tuning.

source

Transducers.usesimd — Method.

usesimd(rf::Reduction, xfsimd::UseSIMD)

Wrap the inner-most loop of reducing function rf with xfsimd. xfsimd is inserted after the inner-most Cat if rf includes Cat.

source

Transducers.@default_finaltype — Macro.

@default_finaltype(xf::Transducer, coll)

Infer the type of the object that would be fed into the second argument input of the bottom reducing function rf(acc, input).

See: Base.@default_eltype

source

Transducers.@next! — Macro.

@next!(rf, state, input)

It is expanded to

state = @next(rf, state, input)

source

Transducers.@simd_if — Macro.

@simd_if rf for ... end

Wrap for-loop with @simd if the outer most transducer of the reducing function rf is UseSIMD.

The next must be called via @next! macro. @simd_if appropriately choose to not use result isa Reduced && return result when SIMD is enabled. It also makes sure that the state returned by next is ignored when ivdep is specified.

source

Transducers.@~ — Macro.

(@~ broadcasting_expression) :: Broadcasted

source