Internals

Internals

Transducers.UseSIMD — Type.

UseSIMD{ivdep}()

Tell the reducible to run the inner reducing function using @simd. The reducible can support it using @simd_if.

source

Transducers._teezip_lens — Function.

_teezip_lens(rf) :: Lens

Return a lens to the .value field of the first "unbalanced" Joiner. A Joiner matched with preceding Splitter would be treated as a regular reducing function node. Thus, reducing function rf must have one more Joiner than Splitter.

source

Transducers.foldl_nocomplete — Method.

foldl_nocomplete(rf, init, coll)

Call __foldl__ without calling complete.

source

Transducers.maybe_usesimd — Method.

maybe_usesimd(xform, simd)

Insert UseSIMD to xform if appropriate.

Arguments

xform::Transducer

simd: false, true, or :ivdep.

Examples

julia> using Transducers
       using Transducers: maybe_usesimd

julia> maybe_usesimd(Map(identity), false)
Map(identity)

julia> maybe_usesimd(Map(identity), true)
Transducers.UseSIMD{false}() |>
    Map(identity)

julia> maybe_usesimd(Cat(), true)
Cat() |>
    Transducers.UseSIMD{false}()

julia> maybe_usesimd(Map(sin) |> Cat() |> Map(cos), :ivdep)
Map(sin) |>
    Cat() |>
    Transducers.UseSIMD{true}() |>
    Map(cos)

julia> maybe_usesimd(Map(sin) |> Cat() |> Map(cos) |> Cat() |> Map(tan), true)
Map(sin) |>
    Cat() |>
    Map(cos) |>
    Cat() |>
    Transducers.UseSIMD{false}() |>
    Map(tan)

source

Transducers.simple_transduce — Method.

simple_transduce(xform, step, init, coll)

Simplified version of transduce. For simple transducers Julia may be able to emit a good code. This function exists only for performance tuning.

source

Transducers.usesimd — Method.

usesimd(xform::Transducer, xfsimd::UseSIMD)

Wrap the inner-most loop of Transducer xform with UseSIMD{ivdep}. It's almost equivalent to UseSIMD{ivdep}() |> xform except that UseSIMD{ivdep}() is inserted after the inner-most Cat if xform includes Cat.

source

Transducers.@simd_if — Macro.

@simd_if rf for ... end

Wrap for-loop with @simd if the outer most transducer of the reducing function rf is UseSIMD.

source