Manual

Transducers and Transducible processes

Base.append!

Base.collect

Base.copy!

Base.foldl

Base.foreach

Base.map!

Base.mapfoldl

Transducers.Zip

Transducers.eduction

Transducers.transduce

Base.Channel

Transducers.Cat

Transducers.Completing

Transducers.Count

Transducers.Dedupe

Transducers.Distinct

Transducers.Drop

Transducers.DropLast

Transducers.DropWhile

Transducers.Filter

Transducers.FlagFirst

Transducers.GetIndex

Transducers.Interpose

Transducers.Iterated

Transducers.Keep

Transducers.Map

Transducers.MapCat

Transducers.MapSplat

Transducers.Merge

Transducers.Partition

Transducers.PartitionBy

Transducers.Replace

Transducers.Scan

Transducers.SetIndex

Transducers.Take

Transducers.TakeLast

Transducers.TakeNth

Transducers.TakeWhile

Transducers.TeeZip

Transducible processes

Base.mapfoldl — Function.

mapfoldl(xf, step, itr; init) :: T
transduce(xf, step, init, itr) :: Union{T, Reduced{T}}

Compose transducer xf with reducing step function step and reduce itr using it.

Warning

transduce differs from mapfoldl as Reduced{T} is returned if the transducer xf or step aborts the reduction.

This API is modeled after transduce in Clojure.

Arguments

xf::Transducer: A transducer.

step: A callable which accepts 1 and 2 arguments. If it only accepts 2 arguments, wrap it by Completing to add complete protocol.

init: An initial value fed to the first argument to reducing step function step.

itr: An iterable.

Examples

julia> using Transducers

julia> function step_demo(state, input)
           @show state, input
           state + input
       end;

julia> function step_demo(state)
           println("Finishing with state = ", state)
           state
       end;

julia> mapfoldl(Filter(isodd), step_demo, 1:4, init=0.0)
(state, input) = (0.0, 1)
(state, input) = (1.0, 3)
Finishing with state = 4.0
4.0

source

Transducers.transduce — Function.

mapfoldl(xf, step, itr; init) :: T
transduce(xf, step, init, itr) :: Union{T, Reduced{T}}

See mapfoldl.

source

Base.foldl — Function.

foldl(step, xf::Transducer, itr; init)
foldl(step, ed::Eduction; init)

The first form is a shorthand for mapfoldl(xf, Completing(step), itr; init). It is intended to be used with do block. It is also equivalent to foldl(step, eduction(xf, itr); init).

Examples

julia> using Transducers

julia> foldl(Filter(isodd), 1:4, init=0.0) do state, input
           @show state, input
           state + input
       end
(state, input) = (0.0, 1)
(state, input) = (1.0, 3)
4.0

source

Base.foreach — Function.

foreach(eff, xf::Transducer, itr)
foreach(eff, ed::Eduction)

Feed the results of xf processing items in itr into a unary function eff which is used primary for a side-effect. It is equivalent to foreach(eff, eduction(xf, coll)). Note that

foreach(eduction(xf, coll)) do x
    ...
end

can be more efficient than

for x in eduction(xf, coll)
    ...
end

as the former does not have to translate the transducer protocol to the iterator protocol.

Examples

julia> using Transducers

julia> foreach(eduction(Filter(isodd), 1:4)) do input
           @show input
       end
input = 1
input = 3

source

Transducers.eduction — Function.

eduction(xf::Transducer, coll)

Create a iterable and reducible object.

Iterable.

Reducible; i.e., it can be handled by transduce efficiently.

This API is modeled after eduction in Clojure.

Examples

julia> using Transducers

julia> for x in eduction(Filter(isodd) |> Take(3), 1:1000)
           @show x
       end
x = 1
x = 3
x = 5

source

Base.map! — Function.

map!(xf::Transducer, dest, src)

Feed src to transducer xf, storing the result in dest. Collections dest and src must have the same shape. Transducer xf may contain filtering transducers. If some entries src are skipped, the corresponding entries in dest will be unchanged. Transducer xf must not contain any expansive transducers such as MapCat.

See also copy!.

Examples

julia> using Transducers

julia> xs = collect(1:5)
       ys = zero(xs)
       map!(Filter(isodd), ys, xs)
5-element Array{Int64,1}:
 1
 0
 3
 0
 5

julia> ans === ys
true

source

Base.copy! — Function.

copy!(xf::Transducer, dest, src)

Feed src to transducer xf, storing the result in dest. Collections dest and src may have the same shape. Source src must be iterable. Destination dest must implement empty! and push!.

See also map!.

Examples

julia> using Transducers

julia> copy!(PartitionBy(x -> x ÷ 3) |> Map(sum), Int[], 1:10)
4-element Array{Int64,1}:
  3
 12
 21
 19

source

Base.append! — Function.

append!(xf::Transducer, dest, src)

This API is modeled after into in Clojure.

Examples

julia> using Transducers

julia> append!(Drop(2), [-1, -2], 1:5)
5-element Array{Int64,1}:
 -1
 -2
  3
  4
  5

source

Base.collect — Function.

collect(xf::Transducer, itr)

Process an iterable itr using a transducer xf and collect the result into a Vector.

Examples

julia> using Transducers

julia> collect(Interpose(missing), 1:3)
5-element Array{Union{Missing, Int64},1}:
 1
  missing
 2
  missing
 3

source

Base.Channel — Type.

Channel(xf::Transducer, itr; kwargs...)
Channel(ed::Eduction; kwargs...)

Pipe items from an iterable itr processed by the transducer xf through a channel. Channel(xf, itr) and Channel(eduction(xf, itr)) are equivalent. Note that itr itself can be a Channel.

Keyword arguments are passed to Channel(function; kwargs...). ctype is inferred from xf if not specified.

Examples

julia> using Transducers

julia> ch1 = Channel(Filter(isodd), 1:5);

julia> ch2 = Channel(Map(x -> 2x - 1), ch1);

julia> ed = eduction(Map(x -> 1:x), ch2);

julia> ch3 = Channel(Cat(), ed);

julia> typeof(ch1) === typeof(ch2) === typeof(ch3) === Channel{Int}
true

julia> foreach(PartitionBy(isequal(1)), ch3) do input
           @show input
       end
input = [1, 1]
input = [2, 3, 4, 5]
input = [1]
input = [2, 3, 4, 5, 6, 7, 8, 9]

source

Transducers

Transducers.Cat — Type.

Cat()

Concatenate/flatten nested iterators.

This API is modeled after cat in Clojure.

Examples

julia> using Transducers

julia> collect(Cat(), [[1, 2], [3], [4, 5]]) == 1:5
true

source

Transducers.Count — Type.

Count([start[, step]])

Generate a sequence start, start + step, start + step + step, and so on.

Note that input is ignored. To use the input in the downstream reduction steps, use TeeZip or Zip.

start defaults to 1 and step defaults to oneunit(start).

See also: Iterators.countfrom.

Examples

julia> using Transducers

julia> collect(TeeZip(Count()), -3:-1)
3-element Array{Tuple{Int64,Int64},1}:
 (-3, 1)
 (-2, 2)
 (-1, 3)

julia> using Dates

julia> collect(TeeZip(Count(Day(1))) |> Map(xs -> *(xs...)), 1:3)
3-element Array{Day,1}:
 1 day
 4 days
 9 days

source

Transducers.Dedupe — Type.

Dedupe()

De-duplicate consecutive items.

This API is modeled after dedupe in Clojure.

Examples

julia> using Transducers

julia> collect(Dedupe(), [1, 1, 2, 1, 3, 3, 2])
5-element Array{Int64,1}:
 1
 2
 1
 3
 2

source

Transducers.Distinct — Type.

Distinct()

Pass only unseen item to the inner reducing step.

This API is modeled after distinct in Clojure.

Examples

julia> using Transducers

julia> collect(Distinct(), [1, 1, 2, 1, 3, 3, 2])
3-element Array{Int64,1}:
 1
 2
 3

source

Transducers.Drop — Type.

Drop(n)

Drop first n items.

This API is modeled after drop in Clojure.

Examples

julia> using Transducers

julia> collect(Drop(3), 1:5)
2-element Array{Int64,1}:
 4
 5

source

Transducers.DropLast — Type.

DropLast(n)

Drop last n items.

This API is modeled after drop-last in Clojure.

Examples

julia> using Transducers

julia> collect(DropLast(2), 1:5)
3-element Array{Int64,1}:
 1
 2
 3

julia> collect(DropLast(2), 1:1)
0-element Array{Int64,1}

julia> collect(DropLast(2), 1:0)
0-element Array{Int64,1}

source

Transducers.DropWhile — Type.

DropWhile(pred)

Drop items while pred returns true consecutively. It becomes a no-op after pred returns a false.

This API is modeled after drop-while in Clojure.

Examples

julia> using Transducers

julia> collect(DropWhile(x -> x < 3), [1:5; 1:2])
5-element Array{Int64,1}:
 3
 4
 5
 1
 2

source

Transducers.Filter — Type.

Filter(pred)

Skip items for which pred is evaluated to false.

This API is modeled after filter in Clojure.

Examples

julia> using Transducers

julia> collect(Filter(iseven), 1:3)
1-element Array{Int64,1}:
 2

source

Transducers.FlagFirst — Type.

FlagFirst()

Output (isfirst, input) where isfirst::Bool is true only for the first iteration and input is the original input.

See also: IterTools.flagfirst

Examples

julia> using Transducers

julia> collect(FlagFirst(), 1:3)
3-element Array{Tuple{Bool,Int64},1}:
 (true, 1)
 (false, 2)
 (false, 3)

source

Transducers.GetIndex — Type.

GetIndex(array)
GetIndex{inbounds}(array)

Transform an integer input i to array[i].

Examples

julia> using Transducers

julia> collect(GetIndex(1:10), [2, 3, 4])
3-element Array{Int64,1}:
 2
 3
 4

julia> collect(GetIndex{true}(1:10), [2, 3, 4])
3-element Array{Int64,1}:
 2
 3
 4

source

Transducers.Interpose — Type.

Interpose(sep)

Interleave input items with a sep.

This API is modeled after interpose in Clojure.

Examples

julia> using Transducers

julia> collect(Interpose(missing), 1:3)
5-element Array{Union{Missing, Int64},1}:
 1
  missing
 2
  missing
 3

source

Transducers.Iterated — Type.

Iterated(f, init[, T::Type])

Generate a sequence init, f(init), f(f(init)), f(f(f(init))), and so on.

Note that input is ignored. To use the input in the downstream reduction steps, use TeeZip or Zip.

Use the third argument T to specify the output type of f.

See also: IterTools.iterated

Examples

julia> using Transducers

julia> collect(Iterated(x -> 2x, 1), 1:5)
5-element Array{Int64,1}:
  1
  2
  4
  8
 16

julia> collect(TeeZip(Iterated(x -> 2x, 1)), 1:5)
5-element Array{Tuple{Int64,Int64},1}:
 (1, 1)
 (2, 2)
 (3, 4)
 (4, 8)
 (5, 16)

source

Transducers.Keep — Type.

Keep(f)

Pass non-nothing output of f to the inner reducing step.

This API is modeled after keep in Clojure.

Examples

julia> using Transducers

julia> xf = Keep() do x
           if x < 3
               x + 1
           end
       end;

julia> collect(xf, 1:5)
2-element Array{Int64,1}:
 2
 3

source

Transducers.Map — Type.

Map(f)

Apply unary function f to each input and pass the result to the inner reducing step.

This API is modeled after map in Clojure.

Examples

julia> using Transducers

julia> collect(Map(x -> 2x), 1:3)
3-element Array{Int64,1}:
 2
 4
 6

source

Transducers.MapCat — Type.

MapCat(f)

Concatenate output of f which is expected to return an iterable.

This API is modeled after mapcat in Clojure.

Examples

julia> using Transducers

julia> collect(MapCat(x -> 1:x), 1:3)
6-element Array{Int64,1}:
 1
 1
 2
 1
 2
 3

source

Transducers.MapSplat — Type.

MapSplat(f)

Like Map(f) but calls f(input...) for each input and then pass the result to the inner reducing step.

Examples

julia> using Transducers

julia> collect(MapSplat(*), zip(1:3, 10:10:30))
3-element Array{Int64,1}:
 10
 40
 90

source

Transducers.Merge — Type.

Merge(iterator)

Merge the output from iterator to the stream processed by the inner reduction step.

Examples

julia> using Transducers

julia> collect(Merge(Iterators.cycle("hello")), 1:8)
8-element Array{Tuple{Int64,Char},1}:
 (1, 'h')
 (2, 'e')
 (3, 'l')
 (4, 'l')
 (5, 'o')
 (6, 'h')
 (7, 'e')
 (8, 'l')

julia> collect(Merge(Iterators.repeated([1 2])), 1:4)
4-element Array{Tuple{Int64,Array{Int64,2}},1}:
 (1, [1 2])
 (2, [1 2])
 (3, [1 2])
 (4, [1 2])

julia> collect(Merge(Iterators.product(1:2, 3:5)), 1:100)
6-element Array{Tuple{Int64,Tuple{Int64,Int64}},1}:
 (1, (1, 3))
 (2, (2, 3))
 (3, (1, 4))
 (4, (2, 4))
 (5, (1, 5))
 (6, (2, 5))

source

Transducers.Partition — Type.

Partition(size, step = size, flush = false)
Partition(size; step = size, flush = false)

Sliding window of width size and interval step.

Warning

The vector passed to the inner reducing function is valid only during its immediate reduction step. It must be reduced immediately or copied.

This API is modeled after partition-all in Clojure.

Examples

julia> using Transducers

julia> collect(Partition(3) |> Map(copy), 1:8)
2-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [4, 5, 6]

julia> collect(Partition(3; flush=true) |> Map(copy), 1:8)
3-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [4, 5, 6]
 [7, 8]

julia> collect(Partition(3; step=1) |> Map(copy), 1:8)
6-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [2, 3, 4]
 [3, 4, 5]
 [4, 5, 6]
 [5, 6, 7]
 [6, 7, 8]

source

Transducers.PartitionBy — Type.

PartitionBy(f)

Group input sequence into chunks in which f returns a same value consecutively.

Warning

The vector passed to the inner reducing function is valid only during its immediate reduction step. It must be reduced immediately or copied.

This API is modeled after partition-by in Clojure.

Examples

julia> using Transducers

julia> collect(PartitionBy(x -> (x + 1) ÷ 3) |> Map(copy), 1:9)
4-element Array{Array{Int64,1},1}:
 [1]
 [2, 3, 4]
 [5, 6, 7]
 [8, 9]

source

Transducers.Replace — Type.

Replace(assoc)

Replace each input with the value in the associative container assoc (e.g., a dictionary, array, string) if it matches with a key/index. Otherwise output the input as-is.

This API is modeled after replace in Clojure.

Examples

julia> using Transducers

julia> collect(Replace(Dict('a' => 'A')), "abc")
3-element Array{Char,1}:
 'A'
 'b'
 'c'

julia> collect(Replace([:a, :b, :c]), 0:4)
5-element Array{Union{Int64, Symbol},1}:
 0
  :a
  :b
  :c
 4

julia> collect(Replace("abc"), 0:4)
5-element Array{Union{Char, Int64},1}:
 0
  'a'
  'b'
  'c'
 4

source

Transducers.Scan — Type.

Scan(f, [init])

Accumulate input with binary function f and pass the accumulated result so far to the inner reduction step.

The inner reducing step receives the sequence y₁, y₂, y₃, ..., yₙ, ... when the sequence x₁, x₂, x₃, ..., xₙ, ... is fed to Scan(f).

y₁ = f(x₁, init)
y₂ = f(x₂, y₁)
y₃ = f(x₃, y₂)
...
yₙ = f(xₙ, yₙ₋₁)

This is a generalized version of the prefix sum also known as cumulative sum, inclusive scan, or scan.

Note that the associativity of f is not required when the transducer is used in a process that gurantee an order, such as mapfoldl.

Unless f is a function with known identity element such as +, *, min, max, and append!, the initial state init must be provided.

Examples

julia> using Transducers

julia> collect(Scan(*), 1:3)
3-element Array{Int64,1}:
 1
 2
 6

julia> collect(Map(x -> x + im) |> Scan(*), 1:3)
3-element Array{Complex{Int64},1}:
 1 + 1im
 1 + 3im
 0 + 10im

julia> collect(Scan(*, 10), 1:3)
3-element Array{Int64,1}:
 10
 20
 60

source

Transducers.SetIndex — Type.

SetIndex(array)
SetIndex{inbounds}(array)

Perform array[i] = v for each input pair (i, v).

Examples

julia> using Transducers

julia> ys = zeros(3);

julia> mapfoldl(SetIndex(ys), first ∘ tuple, [(1, 11.1), (3, 33.3)], init=nothing)

julia> ys
3-element Array{Float64,1}:
 11.1
  0.0
 33.3

source

Transducers.Take — Type.

Take(n)

Take n items from the input sequence.

This API is modeled after take in Clojure.

Examples

julia> using Transducers

julia> collect(Take(2), 1:10)
2-element Array{Int64,1}:
 1
 2

julia> collect(Take(5), 1:2)
2-element Array{Int64,1}:
 1
 2

source

Transducers.TakeLast — Type.

TakeLast(n)

Take last n items from the input sequence.

Examples

julia> using Transducers

julia> collect(TakeLast(2), 1:10)
2-element Array{Int64,1}:
  9
 10

julia> collect(TakeLast(5), 1:2)
2-element Array{Int64,1}:
 1
 2

source

Transducers.TakeNth — Type.

TakeNth(n)

Output every n item to the inner reducing step.

This API is modeled after take-nth in Clojure.

Examples

julia> using Transducers

julia> collect(TakeNth(3), 1:9)
3-element Array{Int64,1}:
 1
 4
 7

source

Transducers.TakeWhile — Type.

TakeWhile(pred)

Take items while pred returns true. Abort the transducible process when pred returns false for the first time.

This API is modeled after take-while in Clojure.

Examples

julia> using Transducers

julia> collect(TakeWhile(x -> x < 3), [1, 2, 3, 1, 2])
2-element Array{Int64,1}:
 1
 2

source

Transducers.TeeZip — Type.

TeeZip(xform::Transducer)

Branch input into two "flows", inject one into xform and then merge the output of xform with the original input.

To illustrate how it works, consider the following usage

xf0 |> TeeZip(xf1) |> xf2

where xf0, xf1, and xf2 are some transducers. Schematically, the output yn from xfn flows as follows:

xf0      xf1                       xf2
---- y0 ------ y1 ---.-- (y0, y1) ----->
      |              |
       `-------------'
    "Tee"          "Zip"

Examples

julia> using Transducers

julia> collect(TeeZip(Filter(isodd) |> Map(x -> x + 1)), 1:5)
3-element Array{Tuple{Int64,Int64},1}:
 (1, 2)
 (3, 4)
 (5, 6)

source

Transducers.Zip — Method.

Zip(xforms...)

Zip outputs of transducers xforms in a tuple and pass it to the inner reduction step.

Warning

Head transducers drive tail transducers. Be careful when using it with transducers other than Map, especially the contractive ones like PartitionBy and the expansive ones like MapCat.

Examples

julia> using Transducers

julia> collect(Zip(Map(identity), Map(x -> 10x), Map(x -> 100x)), 1:3)
3-element Array{Tuple{Int64,Int64,Int64},1}:
 (1, 10, 100)
 (2, 20, 200)
 (3, 30, 300)

source

Miscellaneous

Transducers.Completing — Type.

Completing(function)

Wrap a function to add a no-op complete protocol. Use it when passing a function without 1-argument arity to transduce etc.

This API is modeled after completing in Clojure.

source